Extend ModifyDefaultComponentsEvent to better support changing an items default attribute modifiers.
