+++ src/main/resources/META-INF/accesstransformer.cfg
@@ -344,3 +344,7 @@ public net.minecraft.client.gui.screens.worldselection.EditGameRulesScreen markI
 public net.minecraft.client.gui.screens.worldselection.EditGameRulesScreen clearInvalid(Lnet/minecraft/client/gui/screens/worldselection/EditGameRulesScreen$RuleEntry;)V
 public net.minecraft.client.gui.screens.worldselection.EditGameRulesScreen gameRules
 public net.minecraft.client.gui.screens.worldselection.EditGameRulesScreen$EntryFactory
+
+# For extending for an enhanced builder in the ModifyDefaultComponentsEvent
+protected net.minecraft.core.component.DataComponentPatch$Builder <init>()V
+protected net.minecraft.core.component.DataComponentPatch$Builder map
